Allow NurseConnect signups using WhatsApp
This commit uses the Wassup service to check whether the registrant
already has WhatsApp.

If they do, offer them the choice of whether they want to use WhatsApp
or SMS.

If they don't have WhatsApp, the flow is identical to how it was before.

The Wassup call is quite slow so we use the non-blocking option to warm
the cache before making the blocking call during registration.